requirements.txt
telegram==0.0.1
pyodbc==4.0.32
python-dotenv==0.19.0
python-telegram-bot==13.8

